refactor: Implement Roadmap Refactor Targets
This commit addresses the 4 core refactoring goals outlined in ROADMAP.md:

1. **CSS Cleanup (Table of Contents & Consolidation)**:
   - Restructured `static/style.css` with a unified Table of Contents.
   - Grouped and consolidated redundant variables at the `:root` level.
   - Cleaned up obsolete block comments and structured UI styles logically.

2. **Mobile Media Override Discoverability**:
   - Systematically relocated all orphaned `@media (max-width: ...)` mobile overrides in `style.css` so they immediately follow the desktop rules they modify.
   - Added standard `/* MOBILE OVERRIDE */` comments to prevent "CSS did not move" regression bugs.

3. **Tour Core Helper Extraction**:
   - Created `static/js/tour-core.js` to centralize onboarding tooltip and state logic.
   - Refactored 11 duplicated tour functions across `slashCommands.js`, reducing the file size by over 2400 lines while fully preserving manual custom steps.

4. **Dead Code Pass**:
   - Removed entirely unreferenced backend endpoints (e.g., `cleanup_routes.py`).
   - Deleted stale/legacy JS functions (`addCopyBtn_unused`).
   - Purged dead frontend HTML elements, such as the legacy "Load more" pagination buttons in the Document/Archive libraries that had been superseded by inline scroll loaders.
